feat(arc): add float support

+++ demos/transform/lv_demo_transform.c
@@ -147,7 +147,7 @@ static void arc_event_cb(lv_event_t * e)
 {
     lv_obj_t * arc = lv_event_get_target(e);
 
-    int32_t v = lv_arc_get_angle_end(arc);
+    int32_t v = (int32_t)lv_arc_get_angle_end(arc);
     lv_obj_set_style_transform_rotation(card_to_transform, v * 10, 0);
 }
